Got platform collisions working

+++ src/include/entities/squircle.hpp
@@ -10,7 +10,7 @@ private:
 public:
   Vector2 pos;
   Vector2 vel;
-  int width = 64;
+  float width = 64;
 
   Squircle();
   ~Squircle();
